refactor: hooks와 reducer를 별도 디렉토리로 분리
